+++ src/main/java/com/eaze/model/PaymentOrder.java
@@ -5,6 +5,8 @@ import com.eaze.domian.PaymentOrderStatus;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class PaymentOrder {
@@ -13,7 +15,7 @@ public class PaymentOrder {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private Long amount;
+    private BigDecimal amount;
     private PaymentOrderStatus status;
     private PaymentMethod paymentMethod;
