+++ src/main/java/com/eaze/request/CreateOrderRequest.java
@@ -3,9 +3,11 @@ package com.eaze.request;
 import com.eaze.domian.OrderType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateOrderRequest {
     private String coinId;
-    private double quantity;
+    private BigDecimal quantity;
     private OrderType orderType;
 }
