+++ src/main/java/com/eaze/service/domain/OrderService.java
@@ -6,11 +6,12 @@ import com.eaze.model.Order;
 import com.eaze.model.OrderItem;
 import com.eaze.model.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrderService {
     Order createOrder(User user, OrderItem orderItem, OrderType orderType);
     Order getOrderById(Long orderId) throws Exception;
     List<Order> getAllOrdersOfUser(Long userId);
-    Order processOrder(Coin coin, double quantity, OrderType orderType, User user) throws Exception;
+    Order processOrder(Coin coin, BigDecimal quantity, OrderType orderType, User user) throws Exception;
 }
