+++ src/main/java/com/eaze/service/domain/PaymentService.java
@@ -7,15 +7,17 @@ import com.eaze.response.PaymentResponse;
 import com.razorpay.RazorpayException;
 import com.stripe.exception.StripeException;
 
+import java.math.BigDecimal;
+
 public interface PaymentService {
 
-    PaymentOrder createOrder(User user, Long amount, PaymentMethod paymentMethod);
+    PaymentOrder createOrder(User user, BigDecimal amount, PaymentMethod paymentMethod);
 
     PaymentOrder getPaymentOrderById(Long id) throws Exception;
 
     Boolean proceedPaymentOrder(PaymentOrder paymentOrder, String paymentId) throws RazorpayException;
 
-    PaymentResponse createRazorPaymentLink(User user, Long amount, Long userId);
+    PaymentResponse createRazorPaymentLink(User user, BigDecimal amount, Long userId);
 
-    PaymentResponse createStripePaymentLink(User user, Long amount, Long orderId) throws StripeException;
+    PaymentResponse createStripePaymentLink(User user, BigDecimal amount, Long orderId) throws StripeException;
 }
