+++ src/main/java/com/eaze/service/domain/WithdrawalService.java
@@ -3,11 +3,12 @@ package com.eaze.service.domain;
 import com.eaze.model.User;
 import com.eaze.model.Withdrawal;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface WithdrawalService {
 
-    Withdrawal requestWithdrawal(Long amount, User user);
+    Withdrawal requestWithdrawal(BigDecimal amount, User user);
 
     Withdrawal proceedWithdrawal(Long withdrawalId, boolean accept) throws Exception;
